# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/brothers": {
            "get": {
                "description": "Get data from all Brother records in `Brothers` table",
                "tags": [
                    "Brothers"
                ],
                "summary": "Get all Brothers data",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.Brother"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new Brother record row for `Brothers` table",
                "tags": [
                    "Brothers"
                ],
                "summary": "Create Brother record",
                "parameters": [
                    {
                        "description": "Values for new record",
                        "name": "body_params",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Brother"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    }
                }
            }
        },
        "/api/brothers/count": {
            "get": {
                "description": "Get major distribution counts across all members",
                "tags": [
                    "Brothers"
                ],
                "summary": "Get total Brothers count",
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "integer"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/brothers/majors/count": {
            "get": {
                "description": "Get major distribution counts across all members",
                "tags": [
                    "Brothers"
                ],
                "summary": "Get major counts",
                "responses": {
                    "200": {
                        "description": "desc",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/handlers.GetBrotherStatusCount.SemesterCount"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/brothers/statuses": {
            "get": {
                "description": "Get all status records per brother",
                "tags": [
                    "Brothers"
                ],
                "summary": "Get all status records per brother",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Semester filter",
                        "name": "semester",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.BrotherStatus"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/brothers/statuses/count": {
            "get": {
                "description": "Get status counts for all semesters",
                "tags": [
                    "Brothers"
                ],
                "summary": "Get status counts",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Status filter",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/handlers.GetBrotherStatusCount.SemesterCount"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.GetBrotherStatusCount.SemesterCount": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "semester": {
                    "type": "string"
                }
            }
        },
...
        "models.APIResponse": {
            "description": "JSON response format for all API calls",
            "type": "object",
            "properties": {
                "data": {
                    "description": "Omit if nil"
                },
                "message": {
                    "description": "Omit if nil",
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "models.Brother": {
            "description": "Brother information",
            "type": "object",
            "required": [
                "firstName",
                "lastName",
                "major",
                "rollCall",
                "status"
            ],
            "properties": {
                "badStanding": {
                    "type": "integer"
                },
                "brotherID": {
                    "description": "Primary Key",
                    "type": "integer"
                },
                "className": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "lastName": {
                    "type": "string"
                },
                "major": {
                    "type": "string"
                },
                "phoneNumber": {
                    "type": "string"
                },
                "rollCall": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "models.BrotherStatus": {
            "description": "Brother Status information for a semester",
            "type": "object",
            "properties": {
                "brotherID": {
                    "type": "integer"
                },
                "firstName": {
                    "type": "string"
                },
                "lastName": {
                    "type": "string"
                },
                "major": {
                    "type": "string"
                },
                "rollCall": {
                    "type": "integer"
                },
                "semesterLabel": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
    }
}